Add files via upload
In this project, you will be asked to help a film-producing company. They would like to make sure that films, which they are creating are successful. The company collected a large dataset, which includes information about more than 45'000 films: their description, genre, budget, rating, etc.
Your task will be to suggest and justify a way to measure the success of the film and build a machine learning model, which is capable of predicting this quantity.